Cut batch minting logic
